Intermidiate state of object. PAY ATTENTION ON THE CLASS DIAGRAM, WHEN YOU MAKE METHODS FOR ANY CLASS.
There was a huge amoubt of problems with PlayerBase
