--- core/src/main/java/com/rim/IceField/Eskimo.java
@@ -1,18 +1,0 @@
-package com.rim.IceField;
-
-public class Eskimo extends PlayerBase {
-    private boolean usedIgloo;
-
-    public Eskimo()
-    {
-        super();
-        this.tag = "Eskimo";
-        this.heatLevel = 5;
-
-    }
-
-    @Override
-    public void useSkill(Iceberg ice) {
-        System.out.println("An Igloo has been created!");
-    }
-}
